requirements.txt
python-telegram-bot[job-queue]==20.7
pymongo==4.7.2
python-dotenv==1.0.1
openai>=1.30.0,<2.0.0
